+++ packages/web-components/src/components/global-header/components/global-header/src/components/HybridIpaasHeader/__stories__/HybridIPaasHeader.stories.ts
@@ -168,6 +168,7 @@ export const WithSolis: Story = {
         productName="App Connect"
         productKey="appconnect"
         solisEnvironment="local"
+        solisDevMode
         solisSwitcherEnabled
         solisSidekickEnabled>
       </clabs-global-header-hybrid-ipaas>