+++ packages/web-components/src/components/global-header/components/global-header/src/components/HybridIpaasHeader/__tests__/HybridIpaasHeader.test.ts
@@ -376,15 +376,19 @@ describe('HybridIpaasHeader Component', () => {
         productKey="test-productKey"
         basePath="/base"
         assistMeKey="assist-key"
-        solisSwitcherEnabled="true"></clabs-global-header-hybrid-ipaas>`
+        solisDevMode
+        solisSwitcherEnabled></clabs-global-header-hybrid-ipaas>`
     );
     await waitUntil(
       () => el.headerOptions.capabilityName?.label === 'Test Product',
       'headerOptions were not updated as expected'
     );
-
     expect(el.headerOptions?.solisConfig).to.exist;
     expect(el.headerOptions.solisConfig?.isEnabled).to.be.true;
+    expect(el.headerOptions.solisConfig?.is_prod).to.be.false;
+    expect(el.headerOptions.solisConfig?.backendProxy).to.equal(
+      '/base/hybrid-ipaas/v1/proxies/solis/backend'
+    );
   });
 
   it('should handle solis sidekick rendering', async () => {
@@ -399,7 +403,6 @@ describe('HybridIpaasHeader Component', () => {
       html`<clabs-global-header-hybrid-ipaas
         productName="Test Product"
         productKey="test-productKey"
-        basePath="/base"
         assistMeKey="assist-key"
         solisSidekickEnabled></clabs-global-header-hybrid-ipaas>`
     );
@@ -410,6 +413,9 @@ describe('HybridIpaasHeader Component', () => {
 
     expect(el.headerOptions?.sidekickConfig).to.exist;
     expect(el.headerOptions.sidekickConfig?.isEnabled).to.be.true;
+    expect(el.headerOptions.solisConfig?.backendProxy).to.equal(
+      '/hybrid-ipaas/v1/proxies/solis/backend'
+    );
   });
 
   it('should handle solis rendering in an environment other than local', async () => {
@@ -434,7 +440,6 @@ describe('HybridIpaasHeader Component', () => {
       () => el.headerOptions.capabilityName?.label === 'Test Product',
       'headerOptions were not updated as expected'
     );
-
     expect(el.headerOptions?.sidekickConfig).to.exist;
     expect(el.headerOptions.sidekickConfig?.isEnabled).to.be.true;
     expect(el.headerOptions?.solisConfig).to.exist;