+++ packages/web-components/src/components/global-header/components/global-header/src/types/Header.types.ts
@@ -264,7 +264,7 @@ export type SidekickConfig = {
   context?: string;
   insights_enabled?: boolean;
   chat_enabled?: boolean;
-  reports_enabled?: boolean;
+  overview_enabled?: boolean;
   tell_me_more_enabled?: boolean;
 };
 
@@ -335,6 +335,7 @@ export interface SolisConfig {
   cdn_hostname: string;
   deployment_environment: solisDeploymentEnvironment;
   product_id: string;
+  backendProxy?: string;
 }
 
 export interface NavigationItem {
@@ -356,7 +357,7 @@ export type SidekickInfo = {
   context?: string;
   insights_enabled?: boolean;
   chat_enabled?: boolean;
-  reports_enabled?: boolean;
+  overview_enabled?: boolean;
   tell_me_more_enabled?: boolean;
 };
 
@@ -366,6 +367,7 @@ interface solisWindowConfig {
   deployment_environment: solisDeploymentEnvironment;
   sidekick?: SidekickInfo;
   product_id: string;
+  backend_proxy?: string;
 }
 
 declare global {